cli/requirements.txt
arweave-python-client==0.9.1